refactor(WebexActivityStream): move greeting component on its own file
